move code logic
