+++ bin/custom_multiqc.py
@@ -13,7 +13,6 @@ contain these intermediate results
 
 import argparse
 import logging
-import re
 import sys
 from collections import defaultdict
 from pathlib import Path
@@ -24,14 +23,10 @@ import pandas as pd
 from multiqc.plots import bargraph, heatmap
 from multiqc.types import Anchor
 from utils.constant_variables import (
-    ANNOTATION_NULL_VALUES,
-    ANNOTATION_SEGMENT_KEYS,
-    ANNOTATION_SPECIES_KEYS,
     CONTIG_COMPLETENESS_PCONFIG,
     CONTIG_COMPLETENESS_TOP_N,
     CONTIG_TAXONOMY_PCONFIG,
     CONTIG_TAXONOMY_TOP_N,
-    SEGMENT_SEPARATOR,
     TAXON_OTHER,
     TAXON_UNCLASSIFIED,
     TAXON_UNRECONSTRUCTED,
@@ -507,164 +502,6 @@ def add_n_consensus_clusters_to_mqc(dataframe: pd.DataFrame)-> pd.DataFrame:
     return 0
 
 
-def find_annotation_column(dataframe: pd.DataFrame, keys: List[str]) -> Optional[str]:
-    """
-    Find the `(annotation) ...` column matching the first of `keys` that the annotation database
-    actually provided.
-
-    Which keys exist depends entirely on the fasta headers of `--annotation_db`, so match
-    case-insensitively and ignore separators: `segment`, `Segment`, `segment_name` and
-    `segment name` all answer to the same alias. Columns that are present but empty are skipped,
-    which is what happens when a database declares a field it never fills in.
-    """
-    normalise = lambda text: re.sub(r"[\s_\-]+", "", text).lower()
-    available = {
-        normalise(col.removeprefix("(annotation) ")): col
-        for col in dataframe.columns
-        if col.startswith("(annotation) ")
-    }
-    for key in keys:
-        col = available.get(normalise(key))
-        if col is not None and dataframe[col].notna().any():
-            return col
-    return None
-
-
-def clean_annotation_value(value: object) -> Optional[str]:
-    """
-    Return a stripped annotation value, or None when it is one of the many ways the supported
-    databases spell "not recorded" - Virosaurus `N/A`, BV-BRC `nan`, an empty NCBI field.
-    """
-    if value is None or (isinstance(value, float) and pd.isna(value)):
-        return None
-    text = str(value).strip().strip('"')
-    return None if text.lower() in ANNOTATION_NULL_VALUES else text
-
-
-def normalise_segment(value: object) -> Optional[str]:
-    """
-    Normalise a segment label so the same segment from different databases lands in one column.
-
-    BV-BRC records bare numbers (`4`), GenBank-derived titles spell it out (`segment 4`), and
-    Virosaurus uses names (`HA`, `PB2`, `S`). Numbers become `seg 4`, names are kept as written.
-    """
-    segment = clean_annotation_value(value)
-    if segment is None:
-        return None
-
-    segment = re.sub(r"^(?:segment|seg\.?)\s*", "", segment, flags=re.IGNORECASE).strip()
-    if not segment or segment.lower() in ANNOTATION_NULL_VALUES:
-        return None
-    return f"seg {segment}" if segment.isdigit() else segment
-
-
-def collapse_taxa(dataframe: pd.DataFrame, top_n: int) -> pd.Series:
-    """
-    Return the taxon of every final cluster, keeping only the `top_n` most abundant taxa and
-    collapsing the remainder into a single "Other species" category. Each plot passes its own
-    cap, as a barplot runs out of distinguishable colours long before a heatmap runs out of
-    columns.
-
-    Taxonomy comes from the MMseqs2 annotation results already present in the contig overview
-    table. Clusters whose best hit carries no species, and every cluster of a run without
-    annotation at all (`--skip_consensus_annotation`, `--skip_consensus_qc`), land in
-    "Unclassified" - the plots still show how many clusters each sample yielded and how complete
-    they are, which is the part that does not depend on taxonomy.
-    """
-    species_col = find_annotation_column(dataframe, ANNOTATION_SPECIES_KEYS)
-
-    if species_col is None:
-        return pd.Series(TAXON_UNCLASSIFIED, index=dataframe.index)
-
-    taxa = dataframe[species_col].map(clean_annotation_value).fillna(TAXON_UNCLASSIFIED)
-    keep = set(taxa[taxa != TAXON_UNCLASSIFIED].value_counts().head(top_n).index)
-    return taxa.map(lambda taxon: taxon if taxon in keep or taxon == TAXON_UNCLASSIFIED else TAXON_OTHER)
-
-
-def split_taxa_by_segment(dataframe: pd.DataFrame, taxa: pd.Series) -> pd.Series:
-    """
-    Append the genome segment to the taxon of segmented viruses, so influenza's eight segments
-    get a column each instead of being averaged into one number.
-
-    Only species that actually show more than one segment in this run are split - an unsegmented
-    virus keeps its plain species name rather than gaining a meaningless suffix. The "Other
-    species" and "Unclassified" buckets are aggregates of several species and are never split.
-    """
-    segment_col = find_annotation_column(dataframe, ANNOTATION_SEGMENT_KEYS)
-    if segment_col is None:
-        return taxa
-
-    segments = dataframe[segment_col].map(normalise_segment)
-    if not segments.notna().any():
-        return taxa
-
-    logger.info("Splitting segmented species by column %s", segment_col)
-    splittable = {
-        taxon
-        for taxon, group in segments.groupby(taxa)
-        if taxon not in (TAXON_OTHER, TAXON_UNCLASSIFIED) and group.nunique() > 1
-    }
-    return pd.Series(
-        [
-            f"{taxon} {SEGMENT_SEPARATOR} {segment}" if taxon in splittable and segment else taxon
-            for taxon, segment in zip(taxa, segments)
-        ],
-        index=taxa.index,
-    )
-
-
-def order_taxa(taxa: pd.Series) -> List[str]:
-    """
-    Order taxa by abundance, always ending with the two catch-all categories. Segments of the
-    same species stay together and in natural order (seg 2 before seg 10).
-    """
-    def split_segment(taxon: str) -> Tuple[str, str]:
-        species, _, segment = taxon.partition(f" {SEGMENT_SEPARATOR} ")
-        return species, segment
-
-    def sort_key(taxon: str) -> tuple:
-        species, segment = split_segment(taxon)
-        digits = re.sub(r"\D", "", segment)
-        return (-species_counts.get(species, 0), species, int(digits) if digits else 0, segment)
-
-    species_counts = taxa.map(lambda taxon: split_segment(taxon)[0]).value_counts().to_dict()
-    ranked = sorted(
-        (taxon for taxon in taxa.unique() if taxon not in (TAXON_OTHER, TAXON_UNCLASSIFIED)),
-        key=sort_key,
-    )
-    return ranked + [catch_all for catch_all in (TAXON_OTHER, TAXON_UNCLASSIFIED) if (taxa == catch_all).any()]
-
-
-def estimate_completeness(dataframe: pd.DataFrame) -> Tuple[Optional[pd.Series], str]:
-    """
-    Estimate how complete each final consensus genome is, as a percentage.
-
-    CheckV's own `completeness` estimate is used whenever CheckV ran. Without it we fall back to
-    the fraction of the annotation hit's reference genome (`slen`) that the consensus (`qlen`)
-    spans, discounted by the ambiguous bases QUAST counted - so a 350bp contig of a 7kb genome
-    scores ~5% rather than the ~100% that `100 - % N's` on its own would report.
-
-    Returns (None, "") when neither source is available.
-    """
-    if "(checkv) completeness" in dataframe.columns and dataframe["(checkv) completeness"].notna().any():
-        return pd.to_numeric(dataframe["(checkv) completeness"], errors="coerce"), "CheckV"
-
-    if "(quast) % N's" not in dataframe.columns or not dataframe["(quast) % N's"].notna().any():
-        logger.info("No CheckV or QUAST data available - skipping the completeness heatmap")
-        return None, ""
-
-    called = 1 - pd.to_numeric(dataframe["(quast) % N's"], errors="coerce") / 100
-    if all(col in dataframe.columns for col in ("(annotation) qlen", "(annotation) slen")):
-        qlen = pd.to_numeric(dataframe["(annotation) qlen"], errors="coerce")
-        slen = pd.to_numeric(dataframe["(annotation) slen"], errors="coerce")
-        # Unclassified clusters have no hit and so no reference length. Rather than dropping them
-        # out of the heatmap, fall back to the called fraction alone, which is an upper bound.
-        spanned = (qlen / slen).where(slen > 0).fillna(1.0)
-        return (called * spanned * 100).clip(upper=100), "QUAST proxy"
-
-    return called * 100, "QUAST proxy"
-
-
 def add_contig_taxonomy_to_mqc(dataframe: pd.DataFrame, clusters_summary: pd.DataFrame) -> Optional[int]:
     """
     Add the "Contig clusters" module: how many contig clusters each sample yielded and what they
@@ -698,27 +535,14 @@ def add_contig_taxonomy_barplot(module: "mqc.BaseMultiqcModule", df: pd.DataFram
     this is what the standalone cluster summary barplot used to show separately. The second
     dataset drops that segment for when only the reconstructed genomes are of interest.
     """
-    taxa = collapse_taxa(df, CONTIG_TAXONOMY_TOP_N)
-    counts = df.assign(taxon=taxa).groupby(["sample", "taxon"]).size().unstack(fill_value=0)
-    cats: Dict[str, Dict[str, str]] = {taxon: {"name": taxon} for taxon in order_taxa(taxa)}
-    reconstructed = counts.reindex(columns=list(cats), fill_value=0)
-
-    has_totals = not clusters_summary.empty and {"# Clusters", "# Removed clusters"}.issubset(clusters_summary.columns)
-    if has_totals:
-        # Keep samples that clustered but ended up with no consensus at all - they are the ones
-        # worth spotting, and they would otherwise be missing from the plot entirely.
-        samples = reconstructed.index.union(clusters_summary.index)
-        reconstructed = reconstructed.reindex(samples, fill_value=0)
+    reconstructed, with_removed = summarise_clusters_per_taxon(df, clusters_summary, CONTIG_TAXONOMY_TOP_N)
+    cats: Dict[str, Dict[str, str]] = {taxon: {"name": taxon} for taxon in reconstructed.columns}
 
     datasets = [reconstructed.to_dict(orient="index")]
     cats_per_dataset = [cats]
     data_labels = [{"name": "Reconstructed clusters", "ylab": "# clusters"}]
 
-    if has_totals:
-        total = clusters_summary[["# Clusters", "# Removed clusters"]].sum(axis=1)
-        total = pd.to_numeric(total, errors="coerce").reindex(reconstructed.index).fillna(0)
-        with_removed = reconstructed.copy()
-        with_removed[TAXON_UNRECONSTRUCTED] = (total - reconstructed.sum(axis=1)).clip(lower=0)
+    if with_removed is not None:
         datasets.insert(0, with_removed.to_dict(orient="index"))
         cats_per_dataset.insert(0, {**cats, TAXON_UNRECONSTRUCTED: {"name": TAXON_UNRECONSTRUCTED, "color": "#c8c8c8"}})
         data_labels.insert(0, {"name": "All clusters", "ylab": "# clusters"})
@@ -747,33 +571,16 @@ def add_contig_taxonomy_barplot(module: "mqc.BaseMultiqcModule", df: pd.DataFram
 
 def add_contig_completeness_heatmap(module: "mqc.BaseMultiqcModule", df: pd.DataFrame) -> None:
     """
-    Heatmap of the mean consensus genome completeness per sample and taxon.
+    Heatmap of the consensus genome completeness per sample and taxon.
 
     A heatmap rather than a barplot because completeness is a percentage per sample/taxon pair:
     stacking those percentages would be meaningless, and a blank cell readably means "this
     species was not reconstructed in this sample".
     """
-    completeness, source = estimate_completeness(df)
-    if completeness is None:
+    values, source = summarise_completeness_per_taxon(df, CONTIG_COMPLETENESS_TOP_N)
+    if values is None:
         return
 
-    # Two things are resolved more finely here than in the barplot. More species are named, as
-    # a column costs width rather than a legend colour; and segments get a column each, because
-    # averaging the completeness of influenza's eight segments into one number hides exactly what
-    # you want to see. The barplot answers "which species are in this sample" and stays undivided.
-    segmented = split_taxa_by_segment(df, collapse_taxa(df, CONTIG_COMPLETENESS_TOP_N))
-    values = (
-        df.assign(completeness=completeness, segmented=segmented)
-        .dropna(subset=["completeness"])
-        .groupby(["sample", "segmented"])["completeness"]
-        .median()
-        .round(1)
-        .unstack()
-    )
-    if values.empty:
-        return
-
-    values = values.reindex(columns=[taxon for taxon in order_taxa(segmented) if taxon in values.columns])
     rows = [[None if pd.isna(value) else float(value) for value in row] for _, row in values.iterrows()]
 
     plot = heatmap.plot(
